Apply improved types & architecture to Celo
