+++ libs/ledger-live-common/src/families/celo/broadcast.ts
@@ -1,7 +1,11 @@
+import { AccountBridge } from "@ledgerhq/types-live";
 import { patchOperationWithHash } from "../../operation";
+import { Transaction } from "./types";
 import { celoKit } from "./api/sdk";
 
-const broadcast = async ({ signedOperation: { operation, signature } }) => {
+const broadcast: AccountBridge<Transaction>["broadcast"] = async ({
+  signedOperation: { operation, signature },
+}) => {
   const kit = celoKit();
   const { transactionHash } = await kit.web3.eth.sendSignedTransaction(signature);
   return patchOperationWithHash(operation, transactionHash);
