+++ libs/ledger-live-common/src/families/celo/synchronisation.ts
